+++ ic-stable-structures/src/storage/structures/unbounded.rs
@@ -44,7 +44,7 @@ where
     }
 
     /// Add or replace value associated with `key` in stable memory.
-    pub fn insert(&mut self, key: K, value: V) -> Result<()> {
+    pub fn insert(&mut self, key: &K, value: &V) -> Result<()> {
         let canister_id = ic::id();
 
         // If map for `canister_id` is not initialized, initialize it.
@@ -95,37 +95,12 @@ mod tests {
 
     use ic_exports::{
         ic_kit::{inject::get_context, mock_principals, MockContext},
-        stable_structures::{memory_manager::MemoryId, Storable},
+        stable_structures::memory_manager::MemoryId,
     };
 
-    use crate::{unbounded::SlicedStorable, ChunkSize};
+    use crate::test_utils;
 
     use super::StableUnboundedMap;
-
-    #[derive(Debug, Clone, PartialEq, Eq)]
-    struct StringValue(pub String);
-
-    impl Storable for StringValue {
-        fn to_bytes(&self) -> std::borrow::Cow<[u8]> {
-            self.0.to_bytes()
-        }
-
-        fn from_bytes(bytes: Vec<u8>) -> Self {
-            Self(String::from_bytes(bytes))
-        }
-    }
-
-    impl SlicedStorable for StringValue {
-        fn chunk_size() -> ChunkSize {
-            64
-        }
-    }
-
-    fn str_val(len: usize) -> StringValue {
-        let mut s = String::with_capacity(len);
-        s.extend((0..len).map(|_| 'Q'));
-        StringValue(s)
-    }
 
     #[test]
     fn unbounded_map_works() {
@@ -133,16 +108,16 @@ mod tests {
         let mut map = StableUnboundedMap::new(MemoryId::new(0));
         assert!(map.is_empty());
 
-        let long_str = str_val(50000);
-        let medium_str = str_val(5000);
-        let short_str = str_val(50);
+        let long_str = test_utils::str_val(50000);
+        let medium_str = test_utils::str_val(5000);
+        let short_str = test_utils::str_val(50);
 
-        map.insert(0u32, long_str.clone()).unwrap();
-        map.insert(3u32, medium_str.clone()).unwrap();
-        map.insert(5u32, short_str.clone()).unwrap();
-        assert_eq!(map.get(&0), Some(long_str.clone()));
-        assert_eq!(map.get(&3), Some(medium_str.clone()));
-        assert_eq!(map.get(&5), Some(short_str.clone()));
+        map.insert(&0u32, &long_str).unwrap();
+        map.insert(&3u32, &medium_str).unwrap();
+        map.insert(&5u32, &short_str).unwrap();
+        assert_eq!(map.get(&0).as_ref(), Some(&long_str));
+        assert_eq!(map.get(&3).as_ref(), Some(&medium_str));
+        assert_eq!(map.get(&5).as_ref(), Some(&short_str));
 
         let entries: HashMap<_, _> = map.iter().collect();
         let expected = HashMap::from_iter([(0, long_str), (3, medium_str.clone()), (5, short_str)]);
@@ -161,13 +136,13 @@ mod tests {
 
         let mut map = StableUnboundedMap::new(MemoryId::new(0));
 
-        let long_str = str_val(50000);
-        let medium_str = str_val(5000);
+        let long_str = test_utils::str_val(50000);
+        let medium_str = test_utils::str_val(5000);
 
-        map.insert(0u32, long_str.clone()).unwrap();
+        map.insert(&0u32, &long_str).unwrap();
 
         get_context().update_id(mock_principals::bob());
-        map.insert(3u32, medium_str.clone()).unwrap();
+        map.insert(&3u32, &medium_str).unwrap();
 
         get_context().update_id(mock_principals::alice());
         assert_eq!(map.get(&0), Some(long_str));
